+++ src/main/groovy/com/trevorism/auth/model/Oauth2Tokens.groovy
@@ -1,0 +1,11 @@
+package com.trevorism.auth.model
+
+class Oauth2Tokens {
+
+    String access_token
+    String id_token
+    String token_type
+    int expires_in
+    String refresh_token
+    String scope
+}